+++ models/protobuf/README.md
@@ -7,7 +7,6 @@
 - `protobuf.grammar` - Protobuf syntax definition in Lark format
 - `protobuf_lark.py` - Input parser, creates Protobuf AST
 - `protobuf_ast.py` - Dataclass definitions for a Protobuf AST
-- `protobuf_ast_construction.py` - Define helper functions that build AST nodes
 
 ## How to run
